Issue 34: Move ContactBundle into WebBundle
