+++ app/AppKernel.php
@@ -22,7 +22,6 @@ class AppKernel extends Kernel
             new Khatovar\Bundle\HomepageBundle\KhatovarHomepageBundle(),
             new Khatovar\Bundle\MemberBundle\KhatovarMemberBundle(),
             new Khatovar\Bundle\ExactionBundle\KhatovarExactionBundle(),
-            new Khatovar\Bundle\ContactBundle\KhatovarContactBundle(),
         ];
 
         if (in_array($this->getEnvironment(), ['dev', 'test'], true)) {
