+++ app/config/routing.yml
@@ -1,7 +1,3 @@
-khatovar_contact:
-    resource: '@KhatovarContactBundle/Resources/config/routing.yml'
-    prefix:   /contact
-
 khatovar_exaction:
     resource: '@KhatovarExactionBundle/Resources/config/routing.yml'
     prefix:   /dates
