--- src/Khatovar/Bundle/ContactBundle/Resources/config/form_types.yml
@@ -1,14 +1,0 @@
-services:
-    khatovar_contact.form.type.contact:
-        class: Khatovar\Bundle\ContactBundle\Form\Type\ContactType
-        arguments:
-            - '@khatovar_contact.form.subscriber.add_visit_card'
-        tags:
-            - { name: form.type }
-
-    khatovar_contact.form.type.contact_activation:
-        class: Khatovar\Bundle\ContactBundle\Form\Type\ContactActivationType
-        arguments:
-            - '@doctrine.orm.entity_manager'
-        tags:
-            - { name: form.type }
