--- src/Khatovar/Bundle/ContactBundle/Resources/config/routing.yml
@@ -1,34 +1,0 @@
-khatovar_web_contact:
-     path:     /
-     defaults: { _controller: KhatovarContactBundle:Contact:index }
-
-khatovar_web_contact_show:
-    path:     /{id}/show
-    defaults: { _controller: KhatovarContactBundle:Contact:show }
-
-khatovar_web_contact_list:
-    path:     /list
-    defaults: { _controller: KhatovarContactBundle:Contact:list }
-
-khatovar_web_contact_new:
-    path:     /new
-    defaults: { _controller: KhatovarContactBundle:Contact:new }
-
-khatovar_web_contact_create:
-    path:     /create
-    defaults: { _controller: KhatovarContactBundle:Contact:create }
-    methods:  [POST]
-
-khatovar_web_contact_edit:
-    path:     /{id}/edit
-    defaults: { _controller: KhatovarContactBundle:Contact:edit }
-
-khatovar_web_contact_update:
-    path:     /{id}/update
-    defaults: { _controller: KhatovarContactBundle:Contact:update }
-    methods:  [POST, PUT]
-
-khatovar_web_contact_delete:
-    path:     /{id}/delete
-    defaults: { _controller: KhatovarContactBundle:Contact:delete }
-    methods:  [POST, DELETE]
