+++ src/Khatovar/Bundle/PhotoBundle/Entity/Photo.php
@@ -22,7 +22,7 @@
 namespace Khatovar\Bundle\PhotoBundle\Entity;
 
 use Khatovar\Bundle\WebBundle\Entity\Appearance;
-use Khatovar\Bundle\ContactBundle\Entity\Contact;
+use Khatovar\Bundle\WebBundle\Entity\Contact;
 use Khatovar\Bundle\ExactionBundle\Entity\Exaction;
 use Khatovar\Bundle\HomepageBundle\Entity\Homepage;
 use Khatovar\Bundle\MemberBundle\Entity\Member;
