+++ src/Khatovar/Bundle/WebBundle/Resources/config/handlers.yml
@@ -3,4 +3,4 @@ services:
         class: Khatovar\Bundle\WebBundle\Handler\ActivationHandler
         arguments:
             - '@doctrine.orm.entity_manager'
-            - 'Khatovar\Bundle\ContactBundle\Entity\Contact'
+            - 'Khatovar\Bundle\WebBundle\Entity\Contact'
